Fudge timestamp
For reasons that have no doubt to do with the falsehoods that
programmers believe, if I set a timestamp after midnight GMT, it renders
correctly on my local system but with tomorrow's date in prod. I'm sure
there's a config I can set somewhere to fix that, but who has time for
that? In the meantime, just set a timestamp that doesn't trigger this.