[oklab] Show the more color spaces more betterly
This approach hides the extras by default and then shows them all with
one click. Still a plenty clumsy approach (a button might frankly be
more suitable than <details>), but better than what was there before
which was not discoverable or accessible, and I failed to use it even
with the instructions.