Stack monoid revisited
New blog post describing a practical implementation of the stack monoid.
This commit is a draft and there are still a few TODO items. One is an
illustration showing the in-place monoid reduction; this will be similar
to an existing illustration so I think it's still reads ok.

I also have to do some fixup to the gpu-test link, I haven't pushed the
variant that tests coherent and memoryBarrierBuffer, which is actually
the most important case.