Fix typo in `next-dozen-guis`
I fixed a small typo in your latest post (coincidentally, you were talking about `iced` :grin:).

I thoroughly enjoyed the article, as it's usual with your posts. I agree with all of the advice and I am taking notes for `iced`. The compositor and text rendering sections were specially enlightening.

I am also very excited about `piet-gpu` and can't wait to play with it (and `druid-shell` as well). There are plans for heavy important rendering refactors in `iced` soon and I need to review and reexplore the graphics ecosystem in Rust. As you mentioned, it'd be great to start building some common ground.