[parallel_bezier] Fix error calculation in loop case
This one was interesting. As reported in the following tweet, there were cases where the source curve had a loop and the error calculation significantly underestimated the error for a simple arc.

https://twitter.com/aurimasgasiulis/status/1569241905183047682

The issue is that many rays from the source curve to the candidate failed to intersect the candidate at all. I originally assumed this would basically always happen, so the default error value of 0 was used, and that underestimated the error.

You probably just want to reject the candidate when no rays intersect, but this fix just measures against the endpoints to give you a value in the ballpark of the Frechet distance.