Improve robustness
Apply robustness logic borrowed from main Vello dev: switch to derivatives, tweak derivatives to be nonzero, etc.

Also improve drawing of the Euler spiral in the 180 degree U-turn case; it was drawn as a single cubic segment which is fine for small angle deviations, but makes a colinear case. This drawing is very hacky but is sufficient for visual display.